nspawn: move uid shift/chown() code into shared/
(cherry picked from commit b1fb2d971c810e0bdf9ff0ae567a1c6c230e4e5d)

Related: RHEL-143036
